Mejoras Script Template

+++ Template/oct-template.sh
@@ -7,7 +7,7 @@ Help(){
 }
 
 Config(){
-    echo '{"Source":"gitreposource","SourceBranch":"gitbranchsource","Target":"gitrepotarget","TargetBranch":"gitbranchtarget","Replacements":[{"Search":"SearchText","Replace":"ReplaceText"}, {"Search":"<UserSecretsId>[a-fA-F0-9]\{8\}-[a-fA-F0-9]\{4\}-[a-fA-F0-9]\{4\}-[a-fA-F0-9]\{4\}-[a-fA-F0-9]\{12\}<\/UserSecretsId>","Replace":"<UserSecretsId>MySecret</UserSecretsId>"}]}' | jq . > config.json || echo "Archivo de configuracion generado"
+    echo '{"Source":"gitreposource","SourceBranch":"gitbranchsource","Target":"gitrepotarget","TargetBranch":"gitbranchtarget","Replacements":[{"Search":"SearchText","Replace":"ReplaceText"}]}' | jq . > config.json || echo "Archivo de configuracion generado"
 }
 
 declare -A FLAGS
@@ -20,7 +20,7 @@ while getopts "hivgqps" option; do
       i) Config
          exit;;
       v)
-        echo "Version 1.0.0-alpha"
+        echo "Version 1.0.1-alpha"
         exit;;
      \?) # Invalid option
          echo "Error: Invalid option"
@@ -36,8 +36,6 @@ for row in $( echo "${CONFIG_JSON}" | jq -r '.Replacements[] | "r[\(.Search)]+=\
     eval $(echo ${row}); 
 done
 
-echo "${!r[@]}";
-echo "${r[@]}";
 
 #---------------------------------Colores
 # Reset
@@ -70,38 +68,38 @@ if [[ -v "FLAGS[g]" ]] ; then
     if [[ ! -v "FLAGS[s]" ]]; then
         #Eliminar Conexion con el repositorio del template
         git remote rm origin
-        git remote rm origin
         git remote add origin $TARGET
     fi
 fi
 #---------------------------------Reemplazar nombres  / Carpetas
 for source in "${!r[@]}"; do
     NEW_NAME="${r[$source]}"
-    echo "Repalce Files"
-    for file in $(find ./ -type f -name *${source}* -print)
+    FILES=$( echo "find ./ -type f -name \"*${source}*\" -print")
+    for file in $(eval $FILES)
     do
         FILE="${file##*/}"
         # echo "File Path:        ${file}"
         # echo "File Name:        ${file##*/}"
         # echo "New File Name:    ${FILE/$source/$NEW_NAME}"
         # echo "New File Path:    ${file%/*}/${FILE/$source/$NEW_NAME}"
-        if [[ -v "FLAGS[s]" &&  ! -v "FLAGS[q]" ]] ; then
+        if [[ ! -v "FLAGS[q]" ]] ; then
             printf "${Green}mv ${file} ${file%/*}/${FILE/$source/$NEW_NAME}${Color_Off}"
+            echo
         fi
         if [[ ! -v "FLAGS[s]" ]]; then
             eval $(echo mv ${file} ${file%/*}/${FILE/$source/$NEW_NAME})
         fi
     done
-
-    echo "Repalce Folders"
-    find ./ -type d -name *${source}* -printf '%h\0%d\0%p\n' | sort -t '\0' -nr | awk -F '\0' '{print $3}' | while read folder; do
+    FOLDERS=$( echo "find ./ -type d -name \"*${source}*\" -printf '%h\0%d\0%p\n' | sort -t '\0' -nr | awk -F '\0' '{print \$3}'")
+    eval $FOLDERS | while read folder; do
         FOLDER="${folder##*/}"
         #echo "Folder Path:        ${folder}"
         #echo "Folder Name:        ${folder##*/}"
         #echo "New Folder Name:    ${FOLDER/$source/$NEW_NAME}"
         #echo "New Folder Path:    ${folder%/*}/${FOLDER/$source/$NEW_NAME}"
-        if [[ -v "FLAGS[s]" &&  ! -v "FLAGS[q]" ]] ; then
+        if [[ ! -v "FLAGS[q]" ]] ; then
             printf "${Green}mv ${folder} ${folder%/*}/${FOLDER/$source/$NEW_NAME}${Color_Off}"
+            echo
         fi
         if [[ ! -v "FLAGS[s]" ]]; then
             eval $(echo mv ${folder} ${folder%/*}/${FOLDER/$source/$NEW_NAME}})
@@ -113,7 +111,7 @@ for source in "${!r[@]}"; do
     SED_SCRIPT="s/${source}/${r[$source]}/g"
     for file in $(find ./ -type f -not -path "./.git/*" -and -not -path "./src/frontend/.*" -and -not -path "./src/frontend/node_modules/*" -and -not -path "./tools/*"  -and -not -path "./src/backend/${TEMPLATE_VAR}/bin/*"  -and -not -path "./src/backend/${TEMPLATE_VAR}/obj/*")
     do 
-        if [[ -v "FLAGS[s]" &&  ! -v "FLAGS[q]" ]] ; then
+        if [[  ! -v "FLAGS[q]" ]] ; then
             printf "${Color_Off}at file $file ${Green}\n"
             eval $(echo "sed --quiet '${SED_SCRIPT}p' $file")
         fi
@@ -123,12 +121,25 @@ for source in "${!r[@]}"; do
     done
     printf "${Color_Off}"
 done
+#Reemplazar UUID
+SED_SCRIPT="s/<UserSecretsId>[a-fA-F0-9]\{8\}-[a-fA-F0-9]\{4\}-[a-fA-F0-9]\{4\}-[a-fA-F0-9]\{4\}-[a-fA-F0-9]\{12\}<\/UserSecretsId>/<UserSecretsId>$(uuidgen)<\/UserSecretsId>/g"
+for file in $(find ./ -type f -name '*.csproj' )
+do 
+    if [[ ! -v "FLAGS[q]" ]] ; then
+        printf "${Color_Off}at file $file ${Green}\n"
+        eval $(echo "sed --quiet '${SED_SCRIPT}p' $file")
+    fi
+    if [[ ! -v "FLAGS[s]" ]]; then
+        eval $(echo "sed -i '${SED_SCRIPT}' $file")
+    fi
+done
+printf "${Color_Off}"
 #---------------------------------Push Previus Commits in New Repository
-if [[ ! -v "FLAGS[s]" ]]; then
+if [[ ! -v "FLAGS[s]" && -v "FLAGS[p]" ]]; then
     git push -u origin --all
     git push -u origin --tags
 fi
-if [[ -v "FLAGS[s]" ]]; then
+if [[ -v "FLAGS[s]" && -v "FLAGS[g]" ]]; then
     rm -r ./*
     rm -r ./.*
 fi
